pyproject.toml
[project]
name = "credit_fraud_detector" #TODO: Update to the project name
version = "0.1.0"
description = "Add your description here" #TODO: Update to the project name
authors = [{ name = "Yezer Gonzalez", email = "yezergm@gmail.com" }]
readme = "README.md"
keywords = ['Python', 'template', 'uv']
requires-python = ">=3.12,<3.13"
dependencies = [
    "imblearn>=0.0",
    "ipykernel",
    "ipywidgets>=8.1.7",
    "kagglehub>=0.3.12",
    "make",
    "matplotlib>=3.10.3",
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.19.0",
]

[project.scripts]
"app" = "credit_fraud_detector:main" #TODO: Update to the project name

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_untyped_globals = false
allow_redefinition = true
check_untyped_defs = true
color_output = true
error_summary = true
pretty = true

[tool.ruff]
target-version = "py312" 
line-length = 120
indent-width = 4
fix = true

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
